Splitting up connector code and tests
