+++ src/fides/api/service/connectors/query_configs/google_cloud_postgres_query_config.py
@@ -1,0 +1,7 @@
+from fides.api.service.connectors.query_configs.query_config import (
+    QueryStringWithoutTuplesOverrideQueryConfig,
+)
+
+
+class GoogleCloudSQLPostgresQueryConfig(QueryStringWithoutTuplesOverrideQueryConfig):
+    """Generates SQL in Google Cloud SQL for Postgres' custom dialect."""
