+++ src/fides/api/service/connectors/query_configs/microsoft_sql_server_query_config.py
@@ -1,0 +1,9 @@
+from fides.api.service.connectors.query_configs.query_config import (
+    QueryStringWithoutTuplesOverrideQueryConfig,
+)
+
+
+class MicrosoftSQLServerQueryConfig(QueryStringWithoutTuplesOverrideQueryConfig):
+    """
+    Generates SQL valid for SQLServer.
+    """
