+++ src/fides/api/service/connectors/query_configs/mongodb_query_config.py
@@ -1,0 +1,100 @@
+# pylint: disable=too-many-lines
+from typing import Any, Dict, List, Optional, Tuple
+
+from loguru import logger
+
+from fides.api.models.policy import Policy
+from fides.api.models.privacy_request import PrivacyRequest
+from fides.api.service.connectors.query_configs.query_config import QueryConfig
+from fides.api.util.collection_util import Row, filter_nonempty_values
+
+MongoStatement = Tuple[Dict[str, Any], Dict[str, Any]]
+"""A mongo query is expressed in the form of 2 dicts, the first of which represents
+  the query object(s) and the second of which represents fields to return.
+  e.g. 'collection.find({k1:v1, k2:v2},{f1:1, f2:1 ... })'. This is returned as
+  a tuple ({k1:v1, k2:v2},{f1:1, f2:1 ... }).
+
+  An update statement takes the form
+  collection.update_one({k1:v1},{k2:v2}...}, {$set: {f1:fv1, f2:fv2 ... }}, upsert=False).
+  This is returned as a tuple
+  ({k1:v1},{k2:v2}...},  {f1:fv1, f2: fv2 ... }
+  """
+
+
+class MongoQueryConfig(QueryConfig[MongoStatement]):
+    """Query config that translates parameters into mongo statements"""
+
+    def generate_query(
+        self, input_data: Dict[str, List[Any]], policy: Optional[Policy] = None
+    ) -> Optional[MongoStatement]:
+        def transform_query_pairs(pairs: Dict[str, Any]) -> Dict[str, Any]:
+            """Since we want to do an 'OR' match in mongo, transform queries of the form
+            {A:1, B:2} => "{$or:[{A:1},{B:2}]}".
+            Don't bother to do this if the pairs size is 1
+            """
+            if len(pairs) < 2:
+                return pairs
+            return {"$or": [dict([(k, v)]) for k, v in pairs.items()]}
+
+        if input_data:
+            filtered_data: Dict[str, Any] = self.node.typed_filtered_values(input_data)
+            if filtered_data:
+                query_pairs = {}
+                for string_field_path, data in filtered_data.items():
+                    if len(data) == 1:
+                        query_pairs[string_field_path] = data[0]
+
+                    elif len(data) > 1:
+                        query_pairs[string_field_path] = {"$in": data}
+
+                field_list = {  # Get top-level fields to avoid path collisions
+                    field_path.string_path: 1
+                    for field_path, field in self.top_level_field_map().items()
+                }
+                query_fields, return_fields = (
+                    transform_query_pairs(query_pairs),
+                    field_list,
+                )
+                return query_fields, return_fields
+
+        logger.warning(
+            "There is not enough data to generate a valid query for {}",
+            self.node.address,
+        )
+        return None
+
+    def generate_update_stmt(
+        self, row: Row, policy: Policy, request: PrivacyRequest
+    ) -> Optional[MongoStatement]:
+        """Generate a SQL update statement in the form of Mongo update statement components"""
+        update_clauses = self.update_value_map(row, policy, request)
+
+        pk_clauses: Dict[str, Any] = filter_nonempty_values(
+            {
+                field_path.string_path: field.cast(row[field_path.string_path])
+                for field_path, field in self.primary_key_field_paths.items()
+            }
+        )
+
+        valid = len(pk_clauses) > 0 and len(update_clauses) > 0
+        if not valid:
+            logger.warning(
+                "There is not enough data to generate a valid update for {}",
+                self.node.address,
+            )
+            return None
+        return pk_clauses, {"$set": update_clauses}
+
+    def query_to_str(self, t: MongoStatement, input_data: Dict[str, List[Any]]) -> str:
+        """string representation of a query for logging/dry-run"""
+        query_data, field_list = t
+        db_name = self.node.address.dataset
+        collection_name = self.node.address.collection
+        return f"db.{db_name}.{collection_name}.find({query_data}, {field_list})"
+
+    def dry_run_query(self) -> Optional[str]:
+        data = self.display_query_data()
+        mongo_query = self.generate_query(self.display_query_data(), None)
+        if mongo_query is not None:
+            return self.query_to_str(mongo_query, data)
+        return None
