+++ src/fides/api/service/connectors/query_configs/query_config.py
@@ -1,17 +1,14 @@
 # pylint: disable=too-many-lines
 import re
 from abc import ABC, abstractmethod
-from typing import Any, Dict, Generic, List, Optional, Tuple, Type, TypeVar, Union, cast
+from typing import Any, Dict, Generic, List, Optional, Tuple, Type, TypeVar
 
 import pydash
-from boto3.dynamodb.types import TypeSerializer
-from fideslang.models import MaskingStrategies
 from loguru import logger
 from pydantic import ValidationError
-from sqlalchemy import MetaData, Table, text
-from sqlalchemy.engine import Engine
-from sqlalchemy.sql import Delete, Executable, Update  # type: ignore
-from sqlalchemy.sql.elements import ColumnElement, TextClause
+from sqlalchemy import text
+from sqlalchemy.sql import Executable  # type: ignore
+from sqlalchemy.sql.elements import TextClause
 
 from fides.api.common_exceptions import MissingNamespaceSchemaException
 from fides.api.graph.config import (
@@ -23,10 +20,7 @@ from fides.api.graph.config import (
 )
 from fides.api.graph.execution import ExecutionNode
 from fides.api.models.policy import Policy, Rule
-from fides.api.models.privacy_request import ManualAction, PrivacyRequest
-from fides.api.schemas.namespace_meta.bigquery_namespace_meta import (
-    BigQueryNamespaceMeta,
-)
+from fides.api.models.privacy_request import PrivacyRequest
 from fides.api.schemas.namespace_meta.namespace_meta import NamespaceMeta
 from fides.api.schemas.policy import ActionType
 from fides.api.service.masking.strategy.masking_strategy import MaskingStrategy
@@ -273,86 +267,6 @@ class QueryConfig(Generic[T], ABC):
         """Generate an update statement. If there is no data to be updated
         (for example, if the policy identifies no fields to be updated)
         returns None"""
-
-
-class ManualQueryConfig(QueryConfig[Executable]):
-    def generate_query(
-        self, input_data: Dict[str, List[Any]], policy: Optional[Policy]
-    ) -> Optional[ManualAction]:
-        """Describe the details needed to manually retrieve data from the
-        current collection.
-
-        Example:
-        {
-            "step": "access",
-            "collection": "manual_dataset:manual_collection",
-            "action_needed": [
-                {
-                    "locators": {'email': "customer-1@example.com"},
-                    "get": ["id", "box_id"]
-                    "update":  {}
-                }
-            ]
-        }
-
-        """
-
-        locators: Dict[str, Any] = self.node.typed_filtered_values(input_data)
-        get: List[str] = [
-            field_path.string_path
-            for field_path in self.node.collection.top_level_field_dict
-        ]
-
-        if get and locators:
-            return ManualAction(locators=locators, get=get, update=None)
-        return None
-
-    def query_to_str(self, t: T, input_data: Dict[str, List[Any]]) -> None:
-        """Not used for ManualQueryConfig, we output the dry run query as a dictionary instead of a string"""
-
-    def dry_run_query(self) -> Optional[ManualAction]:  # type: ignore
-        """Displays the ManualAction needed with question marks instead of action data for the locators
-        as a dry run query"""
-        fake_data: Dict[str, Any] = self.display_query_data()
-        manual_query: Optional[ManualAction] = self.generate_query(fake_data, None)
-        if not manual_query:
-            return None
-
-        for where_params in manual_query.locators.values():
-            for i, _ in enumerate(where_params):
-                where_params[i] = "?"
-        return manual_query
-
-    def generate_update_stmt(
-        self, row: Row, policy: Policy, request: PrivacyRequest
-    ) -> Optional[ManualAction]:
-        """Describe the details needed to manually mask data in the
-        current collection.
-
-        Example:
-         {
-            "step": "erasure",
-            "collection": "manual_dataset:manual_collection",
-            "action_needed": [
-                {
-                    "locators": {'id': 1},
-                    "get": []
-                    "update":  {'authorized_user': None}
-                }
-            ]
-        }
-        """
-        locators: Dict[str, Any] = filter_nonempty_values(
-            {
-                field_path.string_path: field.cast(row[field_path.string_path])
-                for field_path, field in self.primary_key_field_paths.items()
-            }
-        )
-        update_stmt: Dict[str, Any] = self.update_value_map(row, policy, request)
-
-        if update_stmt and locators:
-            return ManualAction(locators=locators, get=None, update=update_stmt)
-        return None
 
 
 class SQLLikeQueryConfig(QueryConfig[T], ABC):
@@ -661,36 +575,6 @@ class SQLQueryConfig(SQLLikeQueryConfig[Executable]):
         return f":{query_data_name}"
 
 
-class PostgresQueryConfig(SQLQueryConfig):
-    """
-    Generates SQL valid for Postgres
-    """
-
-    def get_formatted_query_string(
-        self,
-        field_list: str,
-        clauses: List[str],
-    ) -> str:
-        """Returns a query string with double quotation mark formatting for tables that have the same names as
-        Postgres reserved words."""
-        return f'SELECT {field_list} FROM "{self.node.collection.name}" WHERE ({" OR ".join(clauses)})'
-
-
-class MySQLQueryConfig(SQLQueryConfig):
-    """
-    Generates SQL valid for MySQL
-    """
-
-    def get_formatted_query_string(
-        self,
-        field_list: str,
-        clauses: List[str],
-    ) -> str:
-        """Returns a query string with backtick formatting for tables that have the same names as
-        MySQL reserved words."""
-        return f'SELECT {field_list} FROM `{self.node.collection.name}` WHERE ({" OR ".join(clauses)})'
-
-
 class QueryStringWithoutTuplesOverrideQueryConfig(SQLQueryConfig):
     """
     Generates SQL valid for connectors that require the query string to be built without tuples.
@@ -764,425 +648,3 @@ class QueryStringWithoutTuplesOverrideQueryConfig(SQLQueryConfig):
         SELECT order_id,product_id,quantity FROM order_item WHERE order_id IN (:_in_stmt_generated_0, :_in_stmt_generated_1, :_in_stmt_generated_2)
         """
         return self.generate_query_without_tuples(input_data, policy)
-
-
-class MicrosoftSQLServerQueryConfig(QueryStringWithoutTuplesOverrideQueryConfig):
-    """
-    Generates SQL valid for SQLServer.
-    """
-
-
-class SnowflakeQueryConfig(SQLQueryConfig):
-    """Generates SQL in Snowflake's custom dialect."""
-
-    def generate_raw_query(
-        self, field_list: List[str], filters: Dict[str, List[Any]]
-    ) -> Optional[TextClause]:
-        formatted_field_list = [f'"{field}"' for field in field_list]
-        raw_query = super().generate_raw_query(formatted_field_list, filters)
-        return raw_query  # type: ignore
-
-    def format_clause_for_query(
-        self,
-        string_path: str,
-        operator: str,
-        operand: str,
-    ) -> str:
-        """Returns field names in clauses surrounded by quotation marks as required by Snowflake syntax."""
-        return f'"{string_path}" {operator} (:{operand})'
-
-    def get_formatted_query_string(
-        self,
-        field_list: str,
-        clauses: List[str],
-    ) -> str:
-        """Returns a query string with double quotation mark formatting as required by Snowflake syntax."""
-        return f'SELECT {field_list} FROM "{self.node.collection.name}" WHERE ({" OR ".join(clauses)})'
-
-    def format_key_map_for_update_stmt(self, fields: List[str]) -> List[str]:
-        """Adds the appropriate formatting for update statements in this datastore."""
-        fields.sort()
-        return [f'"{k}" = :{k}' for k in fields]
-
-    def get_update_stmt(
-        self,
-        update_clauses: List[str],
-        pk_clauses: List[str],
-    ) -> str:
-        """Returns a parameterised update statement in Snowflake dialect."""
-        return f'UPDATE "{self.node.address.collection}" SET {",".join(update_clauses)} WHERE  {" AND ".join(pk_clauses)}'
-
-
-class RedshiftQueryConfig(SQLQueryConfig):
-    """Generates SQL in Redshift's custom dialect."""
-
-    def get_formatted_query_string(
-        self,
-        field_list: str,
-        clauses: List[str],
-    ) -> str:
-        """Returns a query string with double quotation mark formatting for tables that have the same names as
-        Redshift reserved words."""
-        return f'SELECT {field_list} FROM "{self.node.collection.name}" WHERE ({" OR ".join(clauses)})'
-
-
-class GoogleCloudSQLPostgresQueryConfig(QueryStringWithoutTuplesOverrideQueryConfig):
-    """Generates SQL in Google Cloud SQL for Postgres' custom dialect."""
-
-
-class BigQueryQueryConfig(QueryStringWithoutTuplesOverrideQueryConfig):
-    """
-    Generates SQL valid for BigQuery
-    """
-
-    namespace_meta_schema = BigQueryNamespaceMeta
-
-    @property
-    def partitioning(self) -> Optional[Dict]:
-        # Overriden from base implementation to allow for _only_ BQ partitioning, for now
-        return self.node.collection.partitioning
-
-    def get_partition_clauses(
-        self,
-    ) -> List[str]:
-        """
-        Returns the WHERE clauses specified in the partitioning spec
-
-        Currently, only where-clause based partitioning is supported.
-
-        TODO: derive partitions from a start/end/interval specification
-
-
-        NOTE: when we deprecate `where_clause` partitioning in favor of a more proper partitioning DSL,
-        we should be sure to still support the existing `where_clause` partition definition on
-        any in-progress DSRs so that they can run through to completion.
-        """
-        partition_spec = self.partitioning
-        if not partition_spec:
-            logger.error(
-                "Partitioning clauses cannot be retrieved, no partitioning specification found"
-            )
-            return []
-
-        if where_clauses := partition_spec.get("where_clauses"):
-            return where_clauses
-
-        # TODO: implement more advanced partitioning support!
-
-        raise ValueError(
-            "`where_clauses` must be specified in partitioning specification!"
-        )
-
-    def _generate_table_name(self) -> str:
-        """
-        Prepends the dataset ID and project ID to the base table name
-        if the BigQuery namespace meta is provided.
-        """
-
-        table_name = self.node.collection.name
-        if self.namespace_meta:
-            bigquery_namespace_meta = cast(BigQueryNamespaceMeta, self.namespace_meta)
-            table_name = f"{bigquery_namespace_meta.dataset_id}.{table_name}"
-            if project_id := bigquery_namespace_meta.project_id:
-                table_name = f"{project_id}.{table_name}"
-        return table_name
-
-    def get_formatted_query_string(
-        self,
-        field_list: str,
-        clauses: List[str],
-    ) -> str:
-        """
-        Returns a query string with backtick formatting for tables that have the same names as
-        BigQuery reserved words.
-        """
-        return f'SELECT {field_list} FROM `{self._generate_table_name()}` WHERE ({" OR ".join(clauses)})'
-
-    def generate_masking_stmt(
-        self,
-        node: ExecutionNode,
-        row: Row,
-        policy: Policy,
-        request: PrivacyRequest,
-        client: Engine,
-    ) -> Union[List[Update], List[Delete]]:
-        """
-        Generate a masking statement for BigQuery.
-
-        If a masking override is present, it will take precedence over the policy masking strategy.
-        """
-
-        masking_override = node.collection.masking_strategy_override
-        if masking_override and masking_override.strategy == MaskingStrategies.DELETE:
-            logger.info(
-                f"Masking override detected for collection {node.address.value}: {masking_override.strategy.value}"
-            )
-            return self.generate_delete(row, client)
-        return self.generate_update(row, policy, request, client)
-
-    def generate_update(
-        self, row: Row, policy: Policy, request: PrivacyRequest, client: Engine
-    ) -> List[Update]:
-        """
-        Using TextClause to insert 'None' values into BigQuery throws an exception, so we use update clause instead.
-        Returns a List of SQLAlchemy Update object. Does not actually execute the update object.
-
-        A List of multiple Update objects are returned for partitioned tables; for a non-partitioned table,
-        a single Update object is returned in a List for consistent typing.
-
-        TODO: DRY up this method and `generate_delete` a bit
-        """
-        update_value_map: Dict[str, Any] = self.update_value_map(row, policy, request)
-        non_empty_primary_keys: Dict[str, Field] = filter_nonempty_values(
-            {
-                fpath.string_path: fld.cast(row[fpath.string_path])
-                for fpath, fld in self.primary_key_field_paths.items()
-                if fpath.string_path in row
-            }
-        )
-
-        valid = len(non_empty_primary_keys) > 0 and update_value_map
-        if not valid:
-            logger.warning(
-                "There is not enough data to generate a valid update statement for {}",
-                self.node.address,
-            )
-            return []
-
-        table = Table(self._generate_table_name(), MetaData(bind=client), autoload=True)
-        pk_clauses: List[ColumnElement] = [
-            getattr(table.c, k) == v for k, v in non_empty_primary_keys.items()
-        ]
-
-        if self.partitioning:
-            partition_clauses = self.get_partition_clauses()
-            partitioned_queries = []
-            logger.info(
-                f"Generating {len(partition_clauses)} partition queries for node '{self.node.address}' in DSR execution"
-            )
-            for partition_clause in partition_clauses:
-                partitioned_queries.append(
-                    table.update()
-                    .where(*(pk_clauses + [text(partition_clause)]))
-                    .values(**update_value_map)
-                )
-
-            return partitioned_queries
-
-        return [table.update().where(*pk_clauses).values(**update_value_map)]
-
-    def generate_delete(self, row: Row, client: Engine) -> List[Delete]:
-        """Returns a List of SQLAlchemy DELETE statements for BigQuery. Does not actually execute the delete statement.
-
-        Used when a collection-level masking override is present and the masking strategy is DELETE.
-
-        A List of multiple DELETE statements are returned for partitioned tables; for a non-partitioned table,
-        a single DELETE statement is returned in a List for consistent typing.
-
-        TODO: DRY up this method and `generate_update` a bit
-        """
-
-        non_empty_primary_keys: Dict[str, Field] = filter_nonempty_values(
-            {
-                fpath.string_path: fld.cast(row[fpath.string_path])
-                for fpath, fld in self.primary_key_field_paths.items()
-                if fpath.string_path in row
-            }
-        )
-
-        valid = len(non_empty_primary_keys) > 0
-        if not valid:
-            logger.warning(
-                "There is not enough data to generate a valid DELETE statement for {}",
-                self.node.address,
-            )
-            return []
-
-        table = Table(self._generate_table_name(), MetaData(bind=client), autoload=True)
-        pk_clauses: List[ColumnElement] = [
-            getattr(table.c, k) == v for k, v in non_empty_primary_keys.items()
-        ]
-
-        if self.partitioning:
-            partition_clauses = self.get_partition_clauses()
-            partitioned_queries = []
-            logger.info(
-                f"Generating {len(partition_clauses)} partition queries for node '{self.node.address}' in DSR execution"
-            )
-
-            for partition_clause in partition_clauses:
-                partitioned_queries.append(
-                    table.delete().where(*(pk_clauses + [text(partition_clause)]))
-                )
-
-            return partitioned_queries
-
-        return [table.delete().where(*pk_clauses)]
-
-
-MongoStatement = Tuple[Dict[str, Any], Dict[str, Any]]
-"""A mongo query is expressed in the form of 2 dicts, the first of which represents
-  the query object(s) and the second of which represents fields to return.
-  e.g. 'collection.find({k1:v1, k2:v2},{f1:1, f2:1 ... })'. This is returned as
-  a tuple ({k1:v1, k2:v2},{f1:1, f2:1 ... }).
-
-  An update statement takes the form
-  collection.update_one({k1:v1},{k2:v2}...}, {$set: {f1:fv1, f2:fv2 ... }}, upsert=False).
-  This is returned as a tuple
-  ({k1:v1},{k2:v2}...},  {f1:fv1, f2: fv2 ... }
-  """
-
-
-class MongoQueryConfig(QueryConfig[MongoStatement]):
-    """Query config that translates parameters into mongo statements"""
-
-    def generate_query(
-        self, input_data: Dict[str, List[Any]], policy: Optional[Policy] = None
-    ) -> Optional[MongoStatement]:
-        def transform_query_pairs(pairs: Dict[str, Any]) -> Dict[str, Any]:
-            """Since we want to do an 'OR' match in mongo, transform queries of the form
-            {A:1, B:2} => "{$or:[{A:1},{B:2}]}".
-            Don't bother to do this if the pairs size is 1
-            """
-            if len(pairs) < 2:
-                return pairs
-            return {"$or": [dict([(k, v)]) for k, v in pairs.items()]}
-
-        if input_data:
-            filtered_data: Dict[str, Any] = self.node.typed_filtered_values(input_data)
-            if filtered_data:
-                query_pairs = {}
-                for string_field_path, data in filtered_data.items():
-                    if len(data) == 1:
-                        query_pairs[string_field_path] = data[0]
-
-                    elif len(data) > 1:
-                        query_pairs[string_field_path] = {"$in": data}
-
-                field_list = {  # Get top-level fields to avoid path collisions
-                    field_path.string_path: 1
-                    for field_path, field in self.top_level_field_map().items()
-                }
-                query_fields, return_fields = (
-                    transform_query_pairs(query_pairs),
-                    field_list,
-                )
-                return query_fields, return_fields
-
-        logger.warning(
-            "There is not enough data to generate a valid query for {}",
-            self.node.address,
-        )
-        return None
-
-    def generate_update_stmt(
-        self, row: Row, policy: Policy, request: PrivacyRequest
-    ) -> Optional[MongoStatement]:
-        """Generate a SQL update statement in the form of Mongo update statement components"""
-        update_clauses = self.update_value_map(row, policy, request)
-
-        pk_clauses: Dict[str, Any] = filter_nonempty_values(
-            {
-                field_path.string_path: field.cast(row[field_path.string_path])
-                for field_path, field in self.primary_key_field_paths.items()
-            }
-        )
-
-        valid = len(pk_clauses) > 0 and len(update_clauses) > 0
-        if not valid:
-            logger.warning(
-                "There is not enough data to generate a valid update for {}",
-                self.node.address,
-            )
-            return None
-        return pk_clauses, {"$set": update_clauses}
-
-    def query_to_str(self, t: MongoStatement, input_data: Dict[str, List[Any]]) -> str:
-        """string representation of a query for logging/dry-run"""
-        query_data, field_list = t
-        db_name = self.node.address.dataset
-        collection_name = self.node.address.collection
-        return f"db.{db_name}.{collection_name}.find({query_data}, {field_list})"
-
-    def dry_run_query(self) -> Optional[str]:
-        data = self.display_query_data()
-        mongo_query = self.generate_query(self.display_query_data(), None)
-        if mongo_query is not None:
-            return self.query_to_str(mongo_query, data)
-        return None
-
-
-DynamoDBStatement = Dict[str, Any]
-"""A DynamoDB query is formed using the boto3 library. The required parameters are:
-  * a table/collection name (string)
-  * the key name to pass when accessing the table, along with type and value (dict)
-  * optionally, the sort key or secondary index (i.e. timestamp)
-  * optionally, the specified attributes can be provided. If None, all attributes
-  returned for item.
-
-    # TODO finish these docs
-
-  We can either represent these items as a model and then handle each of the values
-  accordingly in the connector or use this query config to return a dictionary that
-  can be appropriately unpacked when executing using the client.
-
-  The get_item query has been left out of the query_config for now.
-
-  Add an example for put_item
-  """
-
-
-class DynamoDBQueryConfig(QueryConfig[DynamoDBStatement]):
-    def __init__(
-        self, node: ExecutionNode, attribute_definitions: List[Dict[str, Any]]
-    ):
-        super().__init__(node)
-        self.attribute_definitions = attribute_definitions
-
-    def generate_query(
-        self,
-        input_data: Dict[str, List[Any]],
-        policy: Optional[Policy],
-    ) -> Optional[DynamoDBStatement]:
-        """Generates a dictionary for the `query` method used for DynamoDB"""
-        query_param = {}
-        serializer = TypeSerializer()
-        for attribute_definition in self.attribute_definitions:
-            attribute_name = attribute_definition["AttributeName"]
-            attribute_value = input_data[attribute_name][0]
-            query_param["ExpressionAttributeValues"] = {
-                ":value": serializer.serialize(attribute_value)
-            }
-            key_condition_expression: str = f"{attribute_name} = :value"
-            query_param["KeyConditionExpression"] = key_condition_expression  # type: ignore
-        return query_param
-
-    def generate_update_stmt(
-        self, row: Row, policy: Policy, request: PrivacyRequest
-    ) -> Optional[DynamoDBStatement]:
-        """
-        Generate a Dictionary that contains necessary items to
-        run a PUT operation against DynamoDB
-        """
-        update_clauses = self.update_value_map(row, policy, request)
-
-        if update_clauses:
-            serializer = TypeSerializer()
-            update_items = row
-            for key, value in update_items.items():
-                if key in update_clauses:
-                    update_items[key] = serializer.serialize(update_clauses[key])
-                else:
-                    update_items[key] = serializer.serialize(value)
-        else:
-            update_items = None
-
-        return update_items
-
-    def query_to_str(self, t: T, input_data: Dict[str, List[Any]]) -> None:
-        """Not used for this connector"""
-        return None
-
-    def dry_run_query(self) -> None:
-        """Not used for this connector"""
-        return None
