+++ src/fides/api/service/connectors/query_configs/redshift_query_config.py
@@ -1,0 +1,16 @@
+from typing import List
+
+from fides.api.service.connectors.query_configs.query_config import SQLQueryConfig
+
+
+class RedshiftQueryConfig(SQLQueryConfig):
+    """Generates SQL in Redshift's custom dialect."""
+
+    def get_formatted_query_string(
+        self,
+        field_list: str,
+        clauses: List[str],
+    ) -> str:
+        """Returns a query string with double quotation mark formatting for tables that have the same names as
+        Redshift reserved words."""
+        return f'SELECT {field_list} FROM "{self.node.collection.name}" WHERE ({" OR ".join(clauses)})'
