+++ tests/ops/service/connection_config/test_mysql_connector.py
@@ -1,6 +1,6 @@
 from sqlalchemy.orm import Session
 
-from fides.api.service.connectors.sql_connector import MySQLConnector
+from fides.api.service.connectors.mysql_connector import MySQLConnector
 
 
 def test_mysql_connector_build_uri(connection_config_mysql, db: Session):
