+++ tests/ops/service/connectors/test_bigquery_queryconfig.py
@@ -13,7 +13,9 @@ from fides.api.schemas.namespace_meta.bigquery_namespace_meta import (
     BigQueryNamespaceMeta,
 )
 from fides.api.service.connectors import BigQueryConnector
-from fides.api.service.connectors.query_config import BigQueryQueryConfig
+from fides.api.service.connectors.query_configs.bigquery_query_config import (
+    BigQueryQueryConfig,
+)
 
 
 @pytest.mark.integration_external
