+++ tests/ops/service/connectors/test_query_config.py
@@ -21,11 +21,13 @@ from fides.api.models.privacy_request import PrivacyRequest
 from fides.api.schemas.masking.masking_configuration import HashMaskingConfiguration
 from fides.api.schemas.masking.masking_secrets import MaskingSecretCache, SecretType
 from fides.api.schemas.namespace_meta.namespace_meta import NamespaceMeta
-from fides.api.service.connectors.query_config import (
+from fides.api.service.connectors.query_configs.dynamodb_query_config import (
     DynamoDBQueryConfig,
-    MongoQueryConfig,
-    SQLQueryConfig,
 )
+from fides.api.service.connectors.query_configs.mongodb_query_config import (
+    MongoQueryConfig,
+)
+from fides.api.service.connectors.query_configs.query_config import SQLQueryConfig
 from fides.api.service.connectors.scylla_query_config import ScyllaDBQueryConfig
 from fides.api.service.masking.strategy.masking_strategy_hash import HashMaskingStrategy
 from fides.api.util.data_category import DataCategory
